Fix for CF7 4.1
CF7 4.1 introduced a change to the validation functionality in CF7.  The `$result` param is now class instance of the `WPCF7_Validation` class. Created new function to add custom messages and changed format of the invalidation methods.  See http://contactform7.com/2015/01/06/contact-form-7-41-beta/ I have this working in a production environment with no issues so far but I have not run this against any real tests.